+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- An options page to specify the server to target for recording, as well the
+  AppLand server for uploading.
+  
+### Changed
+- Removed AppLand URL from the popup.
+- Update styling
 
 ## [0.1.0] - 2020-05-28
 ### Added